--- src/test/java/models/Address.java
@@ -1,4 +1,0 @@
-package models;
-
-public class Address {
-}
