+++ src/test/java/org/testmonkeys/jentitytest/test/childEntityTests/models/SimpleEntity.java
@@ -1,4 +1,4 @@
-package childEntityTests.models;
+package org.testmonkeys.jentitytest.test.childEntityTests.models;
 
 public class SimpleEntity {
