+++ src/test/java/org/testmonkeys/jentitytest/test/models/Address.java
@@ -1,0 +1,4 @@
+package org.testmonkeys.jentitytest.test.models;
+
+public class Address {
+}
