+++ src/test/java/org/testmonkeys/jentitytest/test/models/Company.java
@@ -1,4 +1,4 @@
-package models;
+package org.testmonkeys.jentitytest.test.models;
 
 import java.util.List;
