+++ src/test/java/org/testmonkeys/jentitytest/test/models/Title.java
@@ -1,4 +1,4 @@
-package models;
+package org.testmonkeys.jentitytest.test.models;
 
 public enum Title {
     Mr,
